+++ packages/dicomImageLoader/src/imageLoader/wadors/metaData/metaDataProvider.ts
@@ -21,9 +21,6 @@ import {
 } from '../../getInstanceModule';
 
 function metaDataProvider(type, imageId) {
-  if (Array.isArray(imageId)) {
-    return;
-  }
   const { MetadataModules } = external.cornerstone.Enums;
   if (type === MetadataModules.MULTIFRAME) {
     // the get function removes the PerFrameFunctionalGroupsSequence
@@ -308,17 +305,18 @@ export function getCineModule(imageId, metaData) {
   const cineRate = getValue<string>(metaData['00180040']);
   return {
     cineRate,
+    numberOfFrames: getNumberValue(metaData['00280008']),
   };
 }
 
 export function getTransferSyntax(imageId, metaData) {
-    // Use either the FMI, which is NOT permitted in the DICOMweb data, but
-    // is sometimes found there anyways, or the available transfer syntax, which
-    // is the recommended way of getting it.
-    return {
-      transferSyntaxUID:
-        getValue<string>(metaData['00020010']) ||
-        getValue<string>(metaData['00083002']),
-    };
+  // Use either the FMI, which is NOT permitted in the DICOMweb data, but
+  // is sometimes found there anyways, or the available transfer syntax, which
+  // is the recommended way of getting it.
+  return {
+    transferSyntaxUID:
+      getValue<string>(metaData['00020010']) ||
+      getValue<string>(metaData['00083002']),
+  };
 }
 export default metaDataProvider;