+++ packages/tools/examples/videoTools/index.ts
@@ -1,4 +1,10 @@
-import { RenderingEngine, Types, Enums } from '@cornerstonejs/core';
+import {
+  RenderingEngine,
+  Types,
+  Enums,
+  eventTarget,
+  triggerEvent,
+} from '@cornerstonejs/core';
 import {
   addButtonToToolbar,
   addDropdownToToolbar,
@@ -15,6 +21,18 @@ console.warn(
 );
 
 const {
+  LengthTool,
+  KeyImageTool,
+  ProbeTool,
+  RectangleROITool,
+  EllipticalROITool,
+  CircleROITool,
+  BidirectionalTool,
+  AngleTool,
+  CobbAngleTool,
+  ArrowAnnotateTool,
+  PlanarFreehandROITool,
+
   PanTool,
   ZoomTool,
   VideoRedactionTool,
@@ -24,8 +42,10 @@ const {
   Enums: csToolsEnums,
 } = cornerstoneTools;
 
+const { annotationFrameRange } = cornerstoneTools.utilities;
+
 const { ViewportType } = Enums;
-const { MouseBindings, KeyboardBindings } = csToolsEnums;
+const { MouseBindings, KeyboardBindings, Events: toolsEvents } = csToolsEnums;
 
 const toolGroupId = 'VIDEO_TOOL_GROUP_ID';
 
@@ -36,8 +56,16 @@ setTitleAndDescription(
 );
 
 const content = document.getElementById('content');
-const element = document.createElement('div');
 
+// Create a selection info element
+const selectionDiv = document.createElement('div');
+selectionDiv.id = 'selection';
+selectionDiv.style.width = '90%';
+selectionDiv.style.height = '1.5em';
+content.appendChild(selectionDiv);
+
+// ************* Create the cornerstone element.
+const element = document.createElement('div');
 // Disable right click context menu so we can have right click tools
 element.oncontextmenu = (e) => e.preventDefault();
 
@@ -60,13 +88,14 @@ rangeElement.oninput = () => {
 };
 
 const instructions = document.createElement('p');
-instructions.innerText = `Playback speed to change CINE playback speed
-Scroll Distance to change amount scrolled on next/prev button or wheel
-Left Drag: Up/down scroll images
-Middle Click or Ctrl+Left: Pan
-Shift+Left: Zoom
-Right Click: Redaction
-Mouse Wheel: Stack Scroll';
+instructions.innerText = `Play/Pause button will toggle the playing of video
+Clear Frame Range clears and selected from range on playback
+Select annotation drop down chooses the tool to use
+Annotation navigation will choose next/previous annotation in the group
+Clicking on the group button switches the displayed annotation group and the group annotations are added to.
+The single image selector sets the annotation to apply to just the current image (shown on +/- 5 frames)
+The [ and ] indicators beside that add left/right boundaries to the image to choose a range.
+Delete annotation will remove an annotation
 `;
 
 content.append(instructions);
@@ -74,91 +103,130 @@ content.append(instructions);
 
 const renderingEngineId = 'myRenderingEngine';
 const viewportId = 'videoViewportId';
+const baseEventDetail = {
+  viewportId,
+  renderingEngineId,
+};
+
 let viewport;
 
-addButtonToToolbar({
+const playButton = addButtonToToolbar({
   id: 'play',
-  title: 'pause',
-  onClick() {
-    viewport.togglePlayPause();
+  title: 'Pause',
+  onClick: (evt) => togglePlay(),
+});
 
-    // toggle the title
-    const button = document.getElementById('play');
-    if (button.innerText === 'pause') {
-      button.innerText = 'play';
-    } else {
-      button.innerText = 'pause';
+const toolsNames = [
+  LengthTool.toolName,
+  KeyImageTool.toolName,
+  ProbeTool.toolName,
+  RectangleROITool.toolName,
+  EllipticalROITool.toolName,
+  CircleROITool.toolName,
+  BidirectionalTool.toolName,
+  AngleTool.toolName,
+  CobbAngleTool.toolName,
+  ArrowAnnotateTool.toolName,
+  PlanarFreehandROITool.toolName,
+  VideoRedactionTool.toolName,
+];
+let selectedToolName = toolsNames[0];
+
+addDropdownToToolbar({
+  options: { values: toolsNames, defaultValue: selectedToolName },
+  onSelectedValueChange: (newSelectedToolNameAsStringOrNumber) => {
+    const newSelectedToolName = String(newSelectedToolNameAsStringOrNumber);
+    const toolGroup = ToolGroupManager.getToolGroup(toolGroupId);
+
+    // Set the new tool active
+    toolGroup.setToolActive(newSelectedToolName, {
+      bindings: [
+        {
+          mouseButton: MouseBindings.Primary, // Left Click
+        },
+      ],
+    });
+
+    // Set the old tool passive
+    toolGroup.setToolPassive(selectedToolName);
+
+    selectedToolName = <string>newSelectedToolName;
+  },
+});
+
+function togglePlay(toggle = undefined) {
+  if (toggle === undefined) {
+    toggle = viewport.togglePlayPause();
+  } else if (toggle === true) {
+    viewport.play();
+  } else {
+    viewport.pause();
+  }
+  playButton.innerText = toggle ? 'Play' : 'Pause';
+}
+
+addButtonToToolbar({
+  id: 'Delete',
+  title: 'Delete Annotation',
+  onClick() {
+    const annotation = getActiveAnnotation();
+    if (annotation) {
+      cornerstoneTools.annotation.state.removeAnnotation(
+        annotation.annotationUID
+      );
     }
   },
 });
 
-addButtonToToolbar({
-  id: 'previous',
-  title: 'previous',
-  onClick() {
-    viewport.scroll(-1);
-  },
-});
+function annotationModifiedListener(evt) {
+  updateAnnotationDiv(
+    evt.detail.annotation?.annotationUID ||
+      evt.detail.annotationUID ||
+      evt.detail.added?.[0]
+  );
+}
 
-addButtonToToolbar({
-  id: 'next',
-  title: 'next',
-  onClick() {
-    viewport.scroll(1);
-  },
-});
+const selectedAnnotation = {
+  annotationUID: '',
+};
 
-addButtonToToolbar({
-  id: 'jump',
-  title: 'jump to 50',
-  onClick() {
-    viewport.setTime(50);
-  },
-});
+function updateAnnotationDiv(uid) {
+  const annotation = cornerstoneTools.annotation.state.getAnnotation(uid);
+  if (!annotation) {
+    selectionDiv.innerHTML = '';
+    selectedAnnotation.annotationUID = '';
+    return;
+  }
+  selectedAnnotation.annotationUID = uid;
+  const { metadata, data } = annotation;
+  const { toolName } = metadata;
+  selectionDiv.innerHTML = `
+    <b>${toolName} Annotation UID:</b>${uid} <b>Label:</b>${
+    data.label || data.text
+  } ${annotation.isVisible ? 'visible' : 'not visible'}
+  `;
+}
 
-const playbackSpeeds = [
-  '0',
-  '0.075',
-  '0.15',
-  '0.25',
-  '0.5',
-  '0.75',
-  '1',
-  '2',
-  '3',
-  '4',
-  '10',
-];
+function getActiveAnnotation() {
+  return cornerstoneTools.annotation.state.getAnnotation(
+    selectedAnnotation.annotationUID
+  );
+}
 
-const toolbar = document.getElementById('demo-toolbar');
-const rateTitle = document.createElement('div');
-rateTitle.style.display = 'inline';
-rateTitle.innerText = 'Playback Rate:';
-toolbar.appendChild(rateTitle);
-addDropdownToToolbar({
-  options: { values: playbackSpeeds, defaultValue: '1', id: 'frameRate' },
-  onSelectedValueChange: (newSelectedToolNameAsStringOrNumber) => {
-    const newPlaybackSpeed = Number(newSelectedToolNameAsStringOrNumber);
-    viewport.setPlaybackRate(newPlaybackSpeed);
-  },
-});
-
-const scrollSpeeds = ['1 f', '2 f', '4 f', '0.5 s', '1 s', '2 s', '4 s'];
-
-const scrollTitle = document.createElement('div');
-scrollTitle.style.display = 'inline';
-scrollTitle.innerText = 'Scroll Distance:';
-toolbar.appendChild(scrollTitle);
-
-addDropdownToToolbar({
-  options: { values: scrollSpeeds, defaultValue: '1 f' },
-  onSelectedValueChange: (value) => {
-    value = value.toString();
-    const unit = value[value.length - 1];
-    const newScrollSpeed = Number(value.substring(0, value.length - 2));
-    viewport.setScrollSpeed(newScrollSpeed, unit);
-  },
-});
+function addAnnotationListeners() {
+  eventTarget.addEventListener(
+    toolsEvents.ANNOTATION_SELECTION_CHANGE,
+    annotationModifiedListener
+  );
+  eventTarget.addEventListener(
+    toolsEvents.ANNOTATION_MODIFIED,
+    annotationModifiedListener
+  );
+  eventTarget.addEventListener(
+    toolsEvents.ANNOTATION_COMPLETED,
+    annotationModifiedListener
+  );
+}
 
 /**
  * Runs the demo
@@ -180,11 +248,26 @@ async function run() {
     (it) => it.indexOf('2.25.179478223177027022014772769075050874231') !== -1
   );
 
+  addAnnotationListeners();
+
+  // Add annotation tools to Cornerstone3D
+  cornerstoneTools.addTool(KeyImageTool);
+  cornerstoneTools.addTool(LengthTool);
+  cornerstoneTools.addTool(ProbeTool);
+  cornerstoneTools.addTool(RectangleROITool);
+  cornerstoneTools.addTool(EllipticalROITool);
+  cornerstoneTools.addTool(CircleROITool);
+  cornerstoneTools.addTool(BidirectionalTool);
+  cornerstoneTools.addTool(AngleTool);
+  cornerstoneTools.addTool(CobbAngleTool);
+  cornerstoneTools.addTool(ArrowAnnotateTool);
+  cornerstoneTools.addTool(PlanarFreehandROITool);
+  cornerstoneTools.addTool(StackScrollMouseWheelTool);
+
   // Add tools to Cornerstone3D
   cornerstoneTools.addTool(PanTool);
   cornerstoneTools.addTool(VideoRedactionTool);
   cornerstoneTools.addTool(ZoomTool);
-  cornerstoneTools.addTool(StackScrollMouseWheelTool);
   cornerstoneTools.addTool(StackScrollTool);
 
   // Define a tool group, which defines how mouse events map to tool commands for
@@ -192,18 +275,40 @@ async function run() {
   const toolGroup = ToolGroupManager.createToolGroup(toolGroupId);
 
   // Add tools to the tool group
+  toolGroup.addTool(LengthTool.toolName);
+  toolGroup.addTool(KeyImageTool.toolName);
+  toolGroup.addTool(ProbeTool.toolName);
+  toolGroup.addTool(RectangleROITool.toolName);
+  toolGroup.addTool(EllipticalROITool.toolName);
+  toolGroup.addTool(CircleROITool.toolName);
+  toolGroup.addTool(BidirectionalTool.toolName);
+  toolGroup.addTool(AngleTool.toolName);
+  toolGroup.addTool(CobbAngleTool.toolName);
+  toolGroup.addTool(ArrowAnnotateTool.toolName);
+  toolGroup.addTool(PlanarFreehandROITool.toolName);
   toolGroup.addTool(PanTool.toolName);
-  toolGroup.addTool(ZoomTool.toolName);
   toolGroup.addTool(VideoRedactionTool.toolName);
+  toolGroup.addTool(StackScrollMouseWheelTool.toolName);
+
+  toolGroup.addTool(ZoomTool.toolName);
   toolGroup.addTool(StackScrollTool.toolName);
 
-  toolGroup.setToolActive(VideoRedactionTool.toolName, {
+  toolGroup.setToolActive(KeyImageTool.toolName, {
     bindings: [
       {
-        mouseButton: MouseBindings.Secondary, // Right Click
+        mouseButton: MouseBindings.Primary,
+        modifierKey: KeyboardBindings.ShiftAlt,
       },
     ],
   });
+  toolGroup.setToolActive(LengthTool.toolName, {
+    bindings: [
+      {
+        mouseButton: MouseBindings.Primary, // Middle Click
+      },
+    ],
+  });
+  toolGroup.setToolActive(StackScrollMouseWheelTool.toolName);
   toolGroup.setToolActive(PanTool.toolName, {
     bindings: [
       {
@@ -223,10 +328,10 @@ async function run() {
       },
     ],
   });
-  toolGroup.setToolActive(StackScrollTool.toolName, {
+  toolGroup.setToolActive(ZoomTool.toolName, {
     bindings: [
       {
-        mouseButton: MouseBindings.Primary, // Left Click
+        mouseButton: MouseBindings.Secondary,
       },
     ],
   });